JS-4200: PopupShortcut as functional component
