+++ src/ts/component/page/elements/head/simple.tsx
@@ -288,21 +288,21 @@ const HeadSimple = observer(class Controls extends React.Component<Props> {
 		const { rootId } = this.props;
 		const object = S.Detail.get(rootId, rootId);
 
-		let sources: string[] = [];
+		let sources: any[] = [];
 
 		switch (object.layout) {
 			case I.ObjectLayout.Type: {
-				sources = S.Record.getTypes().map(it => it.sourceObject);
+				sources = S.Record.getTypes();
 				break;
 			};
 
 			case I.ObjectLayout.Relation: {
-				sources = S.Record.getRelations().map(it => it.sourceObject);
+				sources = S.Record.getRelations();
 				break;
 			};
 		};
 
-		return sources.includes(rootId);
+		return sources.map(it => it.sourceObject).includes(rootId);
 	};
 
 	onCalendar = () => {
